Findbugs "Dodgy code Warnings" addressed (mostly)
I don't know how this one should be fixed:

  Test for floating point equality in mekhq.gui.MekLabTab.refreshRefitSummary()

+++ MekHQ/src/mekhq/campaign/unit/Unit.java
@@ -1173,20 +1173,20 @@ public class Unit implements MekHqXmlSerializable, ITechnology {
                 }
 
                 // fuel tanks
-                partsValue += 200 * js.getFuel() / js.getFuelPerTon();
+                partsValue += 200L * js.getFuel() / js.getFuelPerTon();
 
                 // armor
                 partsValue += js.getArmorWeight(js.locations()) * EquipmentType.getArmorCost(js.getArmorType(0));
 
                 // heat sinks
-                int sinkCost = 2000 + 4000 * js.getHeatType();// == HEAT_DOUBLE ? 6000:
+                long sinkCost = 2000 + 4000l * js.getHeatType();// == HEAT_DOUBLE ? 6000:
                                                            // 2000;
                 partsValue += sinkCost * js.getHeatSinks();
 
                 // grav deck
-                partsValue += 5000000 * js.getGravDeck();
-                partsValue += 10000000 * js.getGravDeckLarge();
-                partsValue += 40000000 * js.getGravDeckHuge();
+                partsValue += 5000000l  * js.getGravDeck();
+                partsValue += 10000000l * js.getGravDeckLarge();
+                partsValue += 40000000l * js.getGravDeckHuge();
 
                 // get bays
                 int baydoors = 0;
